+++ dogapi/dogapi/utils.py
@@ -1,8 +1,12 @@
+from django.conf import settings
+import logging
 from rest_framework.request import Request
 from typing import List
 
 from doglib.pid import pid_factory
 
+logging.config.dictConfig(settings.LOGGING)
+logger = logging.getLogger(__name__)
 
 class QueryparamParsingError(Exception):
     """
@@ -48,7 +52,12 @@ def parse_queryparam(request: Request, param_name: str) -> List[str]:
         query_pid_candidates = request.GET.getlist(f'{param_name}[]')
         if not query_pid_candidates:
             raise QueryparamParsingError(param_name)
-
+    logger.critical(query_pid_candidates)
+    query_pid_candidates = [query_pid_candidate.replace(' ', '') for query_pid_candidate in query_pid_candidates]
     # Swagger UI returns a 1 element list with comma separated values of parameter, e.g. ["string,string,string"]
-    pid_candidates: List[str] = [pid for pid_candidate in query_pid_candidates for pid in pid_candidate.split(',')]
+    pid_candidates: List[str] = [pid.replace(' ', '')
+                                 for pid_candidate in query_pid_candidates
+                                 for pid in pid_candidate.split(',')]
+    logger.critical(pid_candidates)
+
     return pid_candidates